requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.0

requests==2.32.5
openai==2.15.0

llama-index==0.14.12
llama-index-vector-stores-faiss==0.5.1
llama-index-embeddings-nvidia==0.1.5

faiss-cpu==1.7.4
numpy==1.26.4

pdf2image==1.17.0
pillow==10.4.0
opencv-python==4.8.0.74
ultralytics==8.3.235
